JAMES-2521 CLI should not depend on data-library
